Backported malloc and realloc from avr-libc 1.8.0 (without test code) See #857
